Bump SmtpServer from 2.2.0 to 5.3.0
Latest version with .NET standard 1 support

Rebuild IntegrationTests + renames
